* remove ipdb * scipy.misc * modify yaml.load() to yaml.safe_load() * update requirement.txt

+++ datasets/nirscene.py
@@ -1,7 +1,7 @@
 import os
 import random
-import PIL.Image as Image
-from scipy.misc import imread, imsave
+import numpy as np
+from imageio import imread, imsave
 
 
 from gauss_noise import add_gaussian_noise
@@ -71,7 +71,7 @@ def resave(ori_path, tgt_path):
             # add noise
             fn_noise_tgt = os.path.join(tgt_noise_path, foldername + "_" + basename.replace("rgb", "noise") + ".png")
             im_noise = add_gaussian_noise(im, noise_sigma)
-            imsave(fn_noise_tgt, im_noise)
+            imsave(fn_noise_tgt, im_noise.astype(np.uint8))
 
             # nir
             fn_nir = flist_nir[i]
